+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Platynator
 
-## [334](https://github.com/TheMouseNest/Platynator/tree/334) (2026-03-02)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/333...334) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
+## [335](https://github.com/TheMouseNest/Platynator/tree/335) (2026-03-03)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/334...335) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
 
-- Retail: More fixes for mob colours  
+- Cleanup  
+- Use highest alpha of current casting/mouseover on the nameplate  
+- Fix elite colours applying in PvP instances  
+- Add "Scorch" to Execute spells, even though it isn't quite the same  
+- Cleanup  
+- Apply "Interrupt (Not?) Ready" colours to all interrupts on a warlock